+++ lib/mutant/mutator/node/literal/regex.rb
@@ -28,29 +28,15 @@ module Mutant
             emit_type(s(:str, NULL_REGEXP_SOURCE), options)
           end
 
-          # NOTE: will only mutate parts of regexp body if the
-          # body is composed of only strings. Regular expressions
-          # with interpolation are skipped
           def mutate_body
-            return unless body.all?(&method(:n_str?))
+            # NOTE: will only mutate parts of regexp body if the body is composed of only strings.
+            # Regular expressions with interpolation are skipped.
+            return unless (body_ast = AST::Regexp.expand_regexp_ast(input))
 
             Mutator.mutate(body_ast).each do |mutation|
               source = AST::Regexp.to_expression(mutation).to_s
               emit_type(s(:str, source), options)
             end
-          end
-
-          def body_ast
-            AST::Regexp.to_ast(body_expression)
-          end
-
-          def body_expression
-            AST::Regexp.parse(body.map(&:children).join)
-          end
-          memoize :body_expression
-
-          def body
-            children.slice(0...-1)
           end
 
         end # Regex