+++ lib/mutant/mutator/node/send.rb
@@ -32,8 +32,8 @@ module Mutant
           is_a?:         %i[instance_of?],
           kind_of?:      %i[instance_of?],
           map:           %i[each],
-          method:        %i[public_method],
           match:         %i[match?],
+          method:        %i[public_method],
           reverse_each:  %i[each],
           reverse_map:   %i[map each],
           reverse_merge: %i[merge],
@@ -50,6 +50,10 @@ module Mutant
             parse: %i[jd civil strptime iso8601 rfc3339 xmlschema rfc2822 rfc822 httpdate jisx0301]
           }
         )
+
+        REGEXP_MATCH_METHODS    = %i[=~ match match?].freeze
+        REGEXP_START_WITH_NODES = %i[regexp_bos_anchor regexp_literal_literal].freeze
+        REGEXP_END_WITH_NODES   = %i[regexp_literal_literal regexp_eos_anchor].freeze
 
       private
 
@@ -84,6 +88,7 @@ module Mutant
         end
 
         def emit_selector_specific_mutations
+          emit_start_end_with_mutations
           emit_predicate_mutations
           emit_array_mutation
           emit_static_send
@@ -92,6 +97,35 @@ module Mutant
           emit_dig_mutation
           emit_double_negation_mutation
           emit_lambda_mutation
+        end
+
+        def emit_start_end_with_mutations # rubocop:disable Metrics/CyclomaticComplexity, Metrics/MethodLength
+          return unless REGEXP_MATCH_METHODS.include?(selector) && arguments.one?
+
+          argument = Mutant::Util.one(arguments)
+
+          return unless argument.type.equal?(:regexp) && (
+            regexp_ast = AST::Regexp.expand_regexp_ast(argument)
+          )
+
+          regexp_children = regexp_ast.children
+
+          case regexp_children.map(&:type)
+          when REGEXP_START_WITH_NODES
+            emit_start_with(regexp_children)
+          when REGEXP_END_WITH_NODES
+            emit_end_with(regexp_children)
+          end
+        end
+
+        def emit_start_with(regexp_nodes)
+          literal = Mutant::Util.one(regexp_nodes.last.children)
+          emit_type(receiver, :start_with?, s(:str, literal))
+        end
+
+        def emit_end_with(regexp_nodes)
+          literal = Mutant::Util.one(regexp_nodes.first.children)
+          emit_type(receiver, :end_with?, s(:str, literal))
         end
 
         def emit_predicate_mutations