+++ Changelog.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v0.10.26 2021-01-26
 
 * [#1201](https://github.com/mbj/mutant/pull/1201)